LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/zero-column-width-computed-style-expected.txt


PASS column-width:0 is a valid CSS declaration
